requirements.txt

pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
scipy>=1.11
joblib>=1.3
requests>=2.31
jsonschema>=4.19
python-dotenv>=1.0
PyYAML>=6.0
